Change the authentication method for S3 to use AWS SDK authentication (#29)
This diff will change the way gradle authenticates to S3, along with adding some script to handle the assume-role part after getting the federated role. For the federated role login itself, it can't be included here because gradle can't be used (gradle can't run shell executables that requires user interaction), and there is a possibility that whatever program is used for federated login part is broken for the time being (all of them work by scraping the webpage, so whenever Google make some updates to this page it's going to be unusable. Refer to https://29022131.atlassian.net/wiki/spaces/AWS/pages/1187351474/Alternate+way+to+use+AWS+APIs+via+CLI+when+using+Google+SSO for a possible workaround)